fix(color-picker): restore browser interaction coverage
