+++ packages/react/src/components/color-picker/color-picker.test.browser.tsx
@@ -15,6 +15,15 @@ const getInputElement = () => {
   return element
 }
 
+const setInputValue = (input: HTMLInputElement, value: string) => {
+  Object.getOwnPropertyDescriptor(
+    window.HTMLInputElement.prototype,
+    "value",
+  )?.set?.call(input, value)
+
+  input.dispatchEvent(new Event("input", { bubbles: true }))
+}
+
 describe("<ColorPicker />", () => {
   const defaultEyeDropper = (window as any).EyeDropper
   const openEyeDropper = vi.fn()
@@ -60,6 +69,186 @@ describe("<ColorPicker />", () => {
     getInputElement().focus()
 
     await expect.element(getCombobox()).toHaveAttribute("aria-expanded", "true")
+  })
+
+  test("prevents default on mouse down when `openOnFocus` is enabled", async () => {
+    await render(<ColorPicker placeholder="Choose a color" />)
+
+    const event = new MouseEvent("mousedown", {
+      bubbles: true,
+      cancelable: true,
+    })
+
+    getCombobox().element().dispatchEvent(event)
+
+    expect(event.defaultPrevented).toBeTruthy()
+  })
+
+  test("does not prevent default on mouse down when `openOnFocus` is disabled", async () => {
+    await render(
+      <ColorPicker openOnFocus={false} placeholder="Choose a color" />,
+    )
+
+    const event = new MouseEvent("mousedown", {
+      bubbles: true,
+      cancelable: true,
+    })
+
+    getCombobox().element().dispatchEvent(event)
+
+    expect(event.defaultPrevented).toBeFalsy()
+  })
+
+  test("updates input value through `onInputChange` and runs open branch callbacks", async () => {
+    const closeOnChange = vi.fn(() => false)
+    const openOnChange = vi.fn(() => true)
+    const onInputChange = vi.fn()
+
+    await render(
+      <ColorPicker
+        closeOnChange={closeOnChange}
+        openOnChange={openOnChange}
+        placeholder="Choose a color"
+        onInputChange={onInputChange}
+      />,
+    )
+
+    const input = getInputElement()
+
+    setInputValue(input, "#123456")
+
+    expect(onInputChange).toHaveBeenCalledTimes(1)
+    expect(closeOnChange).toHaveBeenCalledTimes(1)
+    expect(openOnChange).toHaveBeenCalledTimes(1)
+    expect(input).toHaveValue("#123456")
+  })
+
+  test("runs close branch callback before open callback on input change", async () => {
+    const closeOnChange = vi.fn(() => true)
+    const openOnChange = vi.fn(() => true)
+
+    await render(
+      <ColorPicker
+        closeOnChange={closeOnChange}
+        openOnChange={openOnChange}
+        placeholder="Choose a color"
+      />,
+    )
+
+    const input = getInputElement()
+
+    setInputValue(input, "#654321")
+
+    expect(closeOnChange).toHaveBeenCalledTimes(1)
+    expect(openOnChange).not.toHaveBeenCalled()
+    expect(input).toHaveValue("#654321")
+  })
+
+  test("ignores input change when `allowInput` is disabled", async () => {
+    const onInputChange = vi.fn()
+
+    await render(
+      <ColorPicker
+        allowInput={false}
+        placeholder="Choose a color"
+        onInputChange={onInputChange}
+      />,
+    )
+
+    const input = getInputElement()
+
+    setInputValue(input, "#ffffff")
+
+    expect(onInputChange).not.toHaveBeenCalled()
+  })
+
+  test("formats and filters typed value using `formatInput` and `pattern`", async () => {
+    await render(
+      <ColorPicker
+        formatInput={(value) => ` ${value} `}
+        pattern={/\s/g}
+        placeholder="Choose a color"
+      />,
+    )
+
+    const input = getInputElement()
+
+    setInputValue(input, "#a1b2c3")
+
+    expect(input).toHaveValue("#a1b2c3")
+  })
+
+  test("keeps invalid value on blur when conversion fails", async () => {
+    await render(<ColorPicker placeholder="Choose a color" />)
+
+    const input = getInputElement()
+
+    setInputValue(input, "invalid-color")
+    input.dispatchEvent(
+      new FocusEvent("blur", {
+        bubbles: true,
+        cancelable: true,
+        relatedTarget: document.body,
+      }),
+    )
+
+    expect(input).toHaveValue("invalid-color")
+  })
+
+  test("keeps value when blur moves focus inside picker field", async () => {
+    await render(<ColorPicker placeholder="Choose a color" />)
+
+    const input = getInputElement()
+
+    setInputValue(input, "#123123")
+    input.dispatchEvent(
+      new FocusEvent("blur", {
+        bubbles: true,
+        cancelable: true,
+        relatedTarget: getCombobox().element(),
+      }),
+    )
+
+    expect(input).toHaveValue("#123123")
+  })
+
+  test("keeps empty value on blur when there is no previous value", async () => {
+    await render(<ColorPicker placeholder="Choose a color" />)
+
+    const input = getInputElement()
+
+    input.dispatchEvent(
+      new FocusEvent("blur", {
+        bubbles: true,
+        cancelable: true,
+        relatedTarget: document.body,
+      }),
+    )
+
+    expect(input).toHaveValue("")
+  })
+
+  test("formats and filters value on blur using `formatInput` and `pattern`", async () => {
+    await render(
+      <ColorPicker
+        defaultValue="#abcdef"
+        formatInput={(value) => `[${value}]`}
+        pattern={/[\[\]]/g}
+        placeholder="Choose a color"
+      />,
+    )
+
+    const input = getInputElement()
+
+    input.dispatchEvent(
+      new FocusEvent("blur", {
+        bubbles: true,
+        cancelable: true,
+        relatedTarget: document.body,
+      }),
+    )
+
+    expect(input).toHaveValue("#abcdef")
   })
 
   test("supports keyboard eye dropper action and writes picked color", async () => {
